Remove FileManager logic to its own file
