+++ tests/unit/cli/test_cli_config_file.py
@@ -234,53 +234,3 @@ class TestCliConfiguration(TestCase):
         get_ctx_defaults("intent-answer", provider, mock_context4, "default")
 
         provider.assert_called_with(None, "default", ["local", "generate-event", "alexa-skills-kit", "intent-answer"])
-
-
-class TestTomlFileManager(TestCase):
-    def test_read_toml(self):
-        config_dir = tempfile.gettempdir()
-        config_path = Path(config_dir, "samconfig.toml")
-        config_path.write_text("version=0.1\n[config_env.topic1.parameters]\nword='clarity'\n")
-        self.assertEqual(
-            TomlFileManager.read(config_path),
-            {"version": 0.1, "config_env": {"topic1": {"parameters": {"word": "clarity"}}}},
-        )
-
-    def test_read_toml_invalid_toml(self):
-        config_dir = tempfile.gettempdir()
-        config_path = Path(config_dir, "samconfig.toml")
-        config_path.write_text("fake='not real'\nimproper toml file\n")
-        with self.assertRaises(FileParseException):
-            TomlFileManager.read(config_path)
-
-    def test_read_toml_file_path_not_valid(self):
-        config_dir = "path/that/doesnt/exist"
-        config_path = Path(config_dir, "samconfig.toml")
-        self.assertEqual(TomlFileManager.read(config_path), {})
-
-    def test_write_toml(self):
-        config_dir = tempfile.gettempdir()
-        config_path = Path(config_dir, "samconfig.toml")
-        toml = {"version": 0.1, "config_env": {"topic2": {"parameters": {"word": "clarity"}}}}
-
-        TomlFileManager.write(toml, config_path)
-
-        txt = config_path.read_text()
-        self.assertIn("version = 0.1", txt)
-        self.assertIn("[config_env.topic2.parameters]", txt)
-        self.assertIn('word = "clarity"', txt)
-
-    def test_dont_write_toml_if_empty(self):
-        config_dir = tempfile.gettempdir()
-        config_path = Path(config_dir, "samconfig.toml")
-        config_path.write_text("nothing to see here\n")
-        toml = {}
-
-        TomlFileManager.write(toml, config_path)
-
-        self.assertEqual(config_path.read_text(), "nothing to see here\n")
-
-    def test_write_toml_bad_path(self):
-        config_path = Path("path/to/some", "file_that_doesnt_exist.toml")
-        with self.assertRaises(FileNotFoundError):
-            TomlFileManager.write({"some key": "some value"}, config_path)
